moved scroll_log and scroll_linear into parameter.utils module
